Removed ScreenMagnifier, Scavenger Hunt, and MARS Xray
All three of the tools have a fair amount of code associated with them.
And ScreenMagnifier and Scavenger Hunt offer next to no value. Xray
offers some value, but is not worth the maintanance of keeping it up to
data.

Other tools may be deleted in the future, but the remaining ones are
either shorter (meaning a lower cost to keep them in) or mush more
useful.