Separate blocking simulate from threaded simulate
Fixes all of the TODOs I made becuase of scary looking code.

The crux of the issue was that in the case of actor == null, the code
would wait for simulation to finish then return with a boolean or throw
an error. In the other case it would just return true.

This leads to some code just ignoreing the return value and the
possiblity of thrown errors because they logically won't be thrown.

Additionally, AbstractMarsToolAndApplication abused (NullPointer)
exceptions in order to acheive stopping behaviour (and error handling
behaviour). That was replaced with an observable (the interface was
already implemented conicidently), and it results in much cleaner code.