Make unterminated strings error and range warnings better
Unterminated strings were mentioned in #29, I decided to make them an
error because it was slightly technically easier. Also I don't know of
any riscv assembler which allows unterminated strings.

The range warnings for .byte and .half were improved to not warn on
values the fit in an unsigned byte/half and to say exactly how the value
is truncated. Now it matches how gcc handles it. This was mentioned in #5