Cleanup as prepared to merge unicode support
Changed the string constructors to use nio.StandardCharset which ensures
that it will not throw an error.

Fixed a bug introduced in SystemIO.writeToFile where the number returned
was not the number of bytes written but the number of characters
written.

Improved the errors generated about \uXXXX in strings

General cleanup of the loop to write bytes in Assembler.storeStrings