Unlock memory and registers when blocked for user IO
This change was made to fix #48

This required changing from using synchronized blocks to explicitly using a
reentrant lock. Everywhere there was a synchronization block there is now a
call to lock() and a try finally with an unlock().

The refactoring I did in the parent was ultimately unneeded as I ended up making
a special unlocker in the blocking methods the system calls use.