Fix bugs related to command line interface
ic, register logging, memory logging and dump sections were all broken by
07581f9e5e733f0e41cfbf0396466dd1af44b3dc . The global state for the program
is now held inside of a program object instead of globally accessible, so
logging gets the zeroed memoery and registers rather than the chnaged ones.

I noticed this while working on documentation for the commandline. The fix is
to pass the program object into the logging functions so they can use local
references to the data rather than reference Globals.memory or RegisterFile.

Dump has not been fixed yet as that is going to touch more files.

Obviously #13 is related, and this also shows how to get lines executed as part
of the api. A bonus of this is that it removes the jank way ic used to be calculated.