Added support for CSR Names in instructions
This makes move.s test finally assemble. It still
doesn't pass, but that should be fixable.

The core idea of this change is to make CSR register
a token type and allow integers to convert to it.
I allowed integer20 to convert despite not technically
being correct. An unsigned 12 bit type would be needed
to make it really accurrate.