Add more options for testing programs
Now comments starting with "error on lines:", "stdout:",
"stdin:", and "stderr:" can now be used to modify how a
test is run and what will make it succeed.

everything after stdin: will be passed to the program as standard in.

stdout and stderr will be compared to the outputs of the
program and cause a failure if they don't match. Notably
the absence of a comment about it means a program should
not output.

If a program should error on assembly it look like:

error on lines: 1, 2, 3, ..., 8