Moved files into a src directory
Per the discussion of restructuring (#33), all of the files that
are included in the jar have been moved intothe src directory.

The build script had to be rewritten, but it should now work with
any files dropped in src. Java files will be compiled over to build
and non-java files will be copied.