Refactored all accesses to memory for .data into a single spot
An alternative assembler could be made that returns a list of
statements and list of datastructs. Maybe space should allocate
a 0 length write at the start and endto make sure that that
gets included in the total length of the data section.

That would allow generating an elf file and therefore allow
more compatability testing.