Add test for self-modifying code and fix bugs it revealed
The old code did not allow incrementally building up an
instruction byte by byte in the text section.

On a different note, the option of switching to nops when
loading a 0 on instruction fetch has been removed as it is
not correct and causes an internal exception rather than the
intended exit codepath.