Fixes the three issues I mentioned on PR #68
For problem 1, the standard code for the help button was used.

For problem 2, a Jlabel was added after the filename input box. It logs
the name of the file is successful and the error message if not.
Unfortunately, it seems it will always reflow onto the next line but not
resize the window making it look like nothing is logged.

For problem 3, changing the layout from a gridbox to a flow did it.