WIP: Fixed the issues in GUI and added config files for memory
The problem with the addresss selector was that it used signed
comparisons rather than unsigned.

I moved the setting of the saved configuration into settings
rather than memoryconfiguration because it prevents circular
static dependency issues.

The memory configuration setter now shows ranges for the various segments.

Config files for memory configurations are now property files. I
will put an example of on the wiki.