Fix program arguments issues
One issue is that if no arguments were passed, the stack pointer
could be pointing outside the stack. 2b15af6f removed the 4K buffer
to the top of the stack, so this bug showed up now.

Similarly it would caused an internal error when run with arguments
because it would try to write to the byte just above the stack.

Additionally, I changed the GUI to actually set the arguments for
blank arguments so you get the desired stack (argc=0,argv={0})

Finally, I cleaned up a little