Allow 64 bit numbers for li and .dword
This was quite difficult to actually implement, but is quite
neccessary to automatically run tests from riscv-tests. Most
of the difficulty stemmed from a bunch of places in the code
expecting the biggest value would be an int rather than a
long.

However, making li work with 64 bit values was also a little
bit of work. It neccessitated making 5 new types of Psuedo-op
templates.

.dword was relatively simple after li wsa working, but required
modifying a significant amount of code.