Add tests from riscv-tests/rv64ui
Significant postproccessing on these tests was needed to get them to
assemble. The test templates use a bunch of shifting and masking which
is not explanded by gcc when outputting assembly. Simpliying the
templates + sed to replace things automatically works pretty well, but
it also uses numbered branches which would force those to be rewrtiten
by hand. Also srl used C shifting logic to generate answers so that
needed to be done by hand.

In general, branches not included and very minor manual changes after
a small script post-processing and template changes.